requirements.txt
flask>=3.0
Pillow>=10.0
numpy>=1.20

# Optional: install from GitHub if you prefer the upstream package
# pip install git+https://github.com/hbldh/hitherdither.git
# (Otherwise the app uses the vendored lib/hitherdither)
